LayoutTests/inspector/console-tests.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="console-tests.js"></script>
<script>

function doit()
{
    console.log('log');
    console.debug('debug');
    console.info('info');
    console.warn('warn');
    console.error('error');
    for (var i = 0; i < 5; ++i)
        console.log('repeated');
    for (var i = 0; i < 2; ++i)
        console.count('count');
    console.group('group');
    console.groupEnd();
    console.log('1', '2', '3');

    dumpConsoleMessagesWithClasses();
}

</script>
</head>

<body onload="onload()">
<p>
Tests that console logging dumps proper messages.
</p>

</body>
</html>
